--- app/src/main/java/com/sbsj/dreamwing/user/MyPointViewModel.kt
@@ -1,4 +1,0 @@
-package com.sbsj.dreamwing.user
-
-class MyPointViewModel {
-}
