--- app/src/main/java/com/sbsj/dreamwing/user/UserActivity.kt
@@ -1,4 +1,0 @@
-package com.sbsj.dreamwing.user
-
-class UserActivity {
-}
