+++ app/src/main/java/com/sbsj/dreamwing/user/model/response/UpdateResponse.kt
@@ -1,5 +1,18 @@
 package com.sbsj.dreamwing.user.model.response
 
+/**
+ * 회원 정보 업데이트 요청에 대한 응답 데이터 클래스
+ *
+ * @author 정은찬
+ * @since 2024.08.05
+ * @version 1.0
+ *
+ * <pre>
+ * 수정일        	수정자        수정내용
+ * ----------  --------    ---------------------------
+ * 2024.08.05   정은찬        최초 생성
+ * </pre>
+ */
 data class UpdateResponse(
     val success: Boolean
 )
