+++ app/src/main/res/layout/activity_my_support_detail.xml
@@ -4,7 +4,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".user.MySupportDetailActivity">
+    tools:context=".user.ui.MySupportDetailActivity">
 
     <include
         android:id="@+id/toolbar"
