+++ app/src/main/res/layout/activity_my_volunteer_detail.xml
@@ -4,7 +4,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".user.MyVolunteerDetailActivity">
+    tools:context=".user.ui.MyVolunteerDetailActivity">
 
     <include
         android:id="@+id/toolbar"
