+++ app/src/main/res/layout/activity_update_user.xml
@@ -137,7 +137,7 @@
             android:textColor="@color/red"
             android:visibility="gone"
             android:layout_marginTop="8dp"
-            android:text="비밀번호가 일치하지 않습니다." />
+            android:text="@string/pwd_no_match" />
 
         <LinearLayout
             android:orientation="vertical"
