+++ app/src/main/res/values/values_eunchan.xml
@@ -3,5 +3,12 @@
     <color name="sub_lightGray1">#F1EFE6</color>
     <color name="sub_lightGray2">#E7E9EA</color>
     <string name="update">회원 정보 수정</string>
-
+    <string name="pwd_no_match">비밀번호가 일치하지 않습니다.</string>
+    <string name="need_login">로그인이 필요합니다.</string>
+    <string name="profile_update">프로필 수정</string>
+    <string name="user_total_support_wing">총 후원 WING</string>
+    <string name="check_details">내역 조회</string>
+    <string name="my_support_details">후원 내역</string>
+    <string name="my_volunteer_details">봉사활동 내역</string>
+    <string name="logout">로그아웃</string>
 </resources>
